backend/requirements.txt
fastapi
uvicorn
pymongo==4.13.2
openai
python-dotenv
pydantic